SKILL-181: satisfy architecture guards for planning-status truth.
Replace constant-returning notStarted() with a const val, move
GoalPlanningStatusAlignRequest into skillbill.application.model, drop
data-class form from the private SharedPreplanSettlement and
GoalPlanningProvenanceRecoverability variants, and allowlist
GoalPlanningRefreshLivenessTest for the SKILL-175 banned-token scan.
