+++ .feature-specs/SKILL-182-cursor-delegated-review/decomposition-manifest.yaml
@@ -1,0 +1,54 @@
+contract_version: "0.5"
+issue_key: "SKILL-182"
+feature_name: "cursor-delegated-review"
+parent_spec_path: ".feature-specs/SKILL-182-cursor-delegated-review/spec.md"
+execution_model: same_branch_commit_per_subtask
+base_branch: main
+feature_branch: feat/SKILL-182-cursor-delegated-review
+stack_branches: []
+current_subtask_intent:
+  subtask_id: 0
+  action: none
+subtasks:
+  - id: 1
+    name: "Cursor harness section in the review-delegation contract"
+    spec_path: ".feature-specs/SKILL-182-cursor-delegated-review/spec_subtask_1_cursor-harness-section.md"
+    status: pending
+    branch: null
+    commit_sha: null
+    workflow_id: null
+    blocked_reason: null
+    last_resumable_step: null
+    linear_issue_id: null
+    dependencies: []
+  - id: 2
+    name: "Render Cursor native agents with Cursor's own frontmatter vocabulary"
+    spec_path: ".feature-specs/SKILL-182-cursor-delegated-review/spec_subtask_2_cursor-frontmatter-render.md"
+    status: pending
+    branch: null
+    commit_sha: null
+    workflow_id: null
+    blocked_reason: null
+    last_resumable_step: null
+    linear_issue_id: null
+    dependencies:
+      - subtask_id: 1
+        optional: false
+        skipped: false
+  - id: 3
+    name: "Reconcile the delegated-review docs with the runtime and the Cursor section"
+    spec_path: ".feature-specs/SKILL-182-cursor-delegated-review/spec_subtask_3_delegated-review-docs.md"
+    status: pending
+    branch: null
+    commit_sha: null
+    workflow_id: null
+    blocked_reason: null
+    last_resumable_step: null
+    linear_issue_id: null
+    dependencies:
+      - subtask_id: 1
+        optional: false
+        skipped: false
+      - subtask_id: 2
+        optional: false
+        skipped: false
