+++ .feature-specs/done/SKILL-178-fix-all-findings-remediation-gate/decomposition-manifest.yaml
