+++ .feature-specs/done/SKILL-178-fix-all-findings-remediation-gate/spec_subtask_1_domain-severity-gates.md
