+++ .feature-specs/done/SKILL-178-fix-all-findings-remediation-gate/spec_subtask_3_human-resumable-block.md
