+++ .feature-specs/done/SKILL-181-preplan-provenance-refresh/decomposition-manifest.yaml
