+++ .feature-specs/done/SKILL-181-preplan-provenance-refresh/spec_subtask_1_validity-freshness-gate.md
